rework architecture variables
Use two spaces instead of tabs. Use one case statement
to declare all required variables for an architecture.
Simpler to recognize any missing or wrong architectures.

Everything is now a test and only one test each run
is allowed. The default is toolchain now. This avoids
unneeded rebuilds of openadk.

There is one open issue, rebuild of uClibc-ng is not detected,
this will be fixed soon, so that testsuite gets installed properly.